Move cut command to its own file
